+++ cl_dll/hl/hl_weapons.cpp
@@ -226,11 +226,11 @@ CBaseEntity::FireBulletsPlayer
 Only produces random numbers to match the server ones.
 =====================
 */
-Vector CBaseEntity::FireBulletsPlayer ( ULONG cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
+Vector CBaseEntity::FireBulletsPlayer (unsigned int cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
 {
 	float x, y, z;
 
-	for ( ULONG iShot = 1; iShot <= cShots; iShot++ )
+	for (unsigned int iShot = 1; iShot <= cShots; iShot++ )
 	{
 		if ( pevAttacker == NULL )
 		{
